+++ src/migrations/index.ts
@@ -51,7 +51,8 @@ import * as migration_20260821_065058_meldungen from './20260821_065058_meldunge
 import * as migration_20260821_213813_abnahme_bausteine_belege from './20260821_213813_abnahme_bausteine_belege';
 import * as migration_20260822_135017_reichweite_sterne_faq from './20260822_135017_reichweite_sterne_faq';
 import * as migration_20260823_073210_google_rezensionen from './20260823_073210_google_rezensionen'
-import * as migration_20260823_110500_meldungen_github from './20260823_110500_meldungen_github';
+import * as migration_20260823_110500_meldungen_github from './20260823_110500_meldungen_github'
+import * as migration_20260823_115500_digitale_ware from './20260823_115500_digitale_ware';
 
 export const migrations = [
   {
@@ -323,5 +324,10 @@ export const migrations = [
     up: migration_20260823_110500_meldungen_github.up,
     down: migration_20260823_110500_meldungen_github.down,
     name: '20260823_110500_meldungen_github'
+  },
+  {
+    up: migration_20260823_115500_digitale_ware.up,
+    down: migration_20260823_115500_digitale_ware.down,
+    name: '20260823_115500_digitale_ware'
   },
 ];